+++ src/test/java/at/fh/burgenland/coordinatesystem/CoordinateSystemControllerTest.java
@@ -1,11 +1,10 @@
-package at.fh.burgenland.coordinatesystemtest;
+package at.fh.burgenland.coordinatesystem;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import at.fh.burgenland.coordinatesystem.CoordinateSystemController;
 import at.fh.burgenland.fft.FrequenzDbOutput;
 import java.lang.reflect.Field;
 import javafx.scene.canvas.Canvas;
