+++ src/test/java/at/fh/burgenland/games/voicezone/VoiceZoneControllerTest.java
@@ -1,11 +1,9 @@
-package at.fh.burgenland.gamestest.voicezonetest;
+package at.fh.burgenland.games.voicezone;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import at.fh.burgenland.games.voicezone.VoiceZoneController;
-import at.fh.burgenland.games.voicezone.VoiceZoneTrainingMode;
 import javafx.application.Platform;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
